+++ game_patch/purefaction/pf.h
@@ -12,3 +12,4 @@ bool pf_process_raw_unreliable_packet(const void* data, int len, const rf::NetAd
 void pf_player_init(rf::Player* player);
 void pf_player_level_load(rf::Player* player);
 bool pf_is_player_verified(rf::Player* player);
+int pf_get_player_ac_level(rf::Player*);