+++ docs/CHANGELOG.md
@@ -12,7 +12,7 @@ Version 1.6.2 (not released yet)
 - Add `pow2_tex` command for forcing usage of power of two textures
 - Add support for redirecting dedicated server standard output stream to a file in win32-console mode
 - Add server-side part of Pure Faction client verification
-- Add `$DF Require Verified Client` option in dedicated server config
+- Add `$DF Anticheat Level` option in dedicated server config
 - Send Kills/Deaths statistics to players joining an already started gameplay
 - Send player statistics (kills, deaths, accuracy) in private message after round ends in Dash Faction server
 - Update fpgun skin when switching players in spectate mode